move schema into separate files
